+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [2.0.4] 2024-11-15
+
+- Add `LONGPORT_PRINT_QUOTE_PACKAGES` environment variable to enable printing the opened quote packages when connected to the server, default is `true`.
+
 # [2.0.3] 2024-11-14
 
 - Changed the `time` parameter of `Quote.history_candlesticks_by_offset` method to be optional.